qa: Postgres Dockerfile for tests
When creating a new repo, please run
cp test/postgresql/postgresql.conf.backup test/postgresql/postgresql.conf

This is needed for testing because when docker container runs, it chowns postgresql.conf to docker, and you no longer have access to it.
You can run "chmod 1000:1000 postgresql.conf" to access the file after the container stops.
But anyway, postgresql.conf is being ignored because of that